requirements.txt
python-telegram-bot
google-cloud-dialogflow